FIN_APP_MAIN/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
requests==2.26.0
python-dotenv==0.19.0
aiohttp==3.8.1